Add files via upload
This document contains the Chicago
Crimes dataset from 2001 to present. Due to upload limitations of the git repository, only  14.8MB  of the 1.4GB was uploaded .
The field contained in the document are as below :-
1. ID : This field reflects the unique identifier for and
individual crime. This is denoted as a seven digit number
.
2.  Case Number : This attribute reflects the case number
for the crime. This number is assigned by a responsible
officer to a docket. All subsequent actions around the
crime incident are then recorded in the opende docket
[?] . The case number for the Chicago Crimes is stored
as a string with a prefix of two letters followed by six
digits.
3. Date : This field reflects the day and time at which the
crime got reported.
4.  Block : This field denotes an the block within the city of
Chicago where the crime occured. The way this field has
been used is such that the exact location of the crime is
kept anonymous to the general public. This field contains
three pieces of information:
– bock unique (e.g. 013XX ) identifier
– Compass Location of crime scene relative to the
block ( e.g. W )
– Street name ( e.g. NORTH AVE )
Combined, the above gives 013XX W NORTH AVE
5. IUCR : This field is used to denote the type of the
commited crime . A four digit numerical code is used
for this.
6. Primary Type : This attribute referes to the catergory of
the crime from known crime classes. THis is stored as a
string ( e.g THEFT ) .

7. Description : This field is a brief description of the crime
commited.

 Location Description : This field gives a brief description
of the area where the crime occured. A string is used for
this .

8. Arrest : This field denotes whether there was an arrest or
not for the crime of concern. This is a boolean ( true/false
) field )

9. Domestic : Indicates whether the crime was domestic or
not. This is a boolean ( true/false ) field )

10. Beat : This field denotes the smallest police area where
the crime occured.

11. District : This field reflect the police district where the
crime occured.

12.  Ward : This attribute represents a city ward where the
crime occured.

13. Community Area : This attribute represents the community
where the crime occured.

14. FBI Code : This field shows the classification of the crime
according to the FBI standards .

15. X Coordinate : This denotes the x coordinate of the
occurance of the crime using the State Plane Illinois East
NAD 1983 coordinate system.

15. Y Coordinate : This denotes the y coordinate of the
occurance of the crime using the State Plane Illinois East
NAD 1983 coordinate system.

16. Year : This field denotes the year at which the crime
occured. This is represented as a four digit number .

17.  Updated On : This field referes to the time when the
record was last updated .

18.  Latitude : This is the geographic latitute of the place
where the crime occured.

19.  Longitude : This is the geographic longitude of the place
where the crime occured.

20. Location : The geographic location of the crime scene
using the longitude and latitude.